Move SVG version of circle info to new component
